+++ packages/graphql/errorLink.ts
@@ -1,29 +1,27 @@
 import { onError } from '@apollo/client/link/error'
 
 export const errorLink = onError(({ graphQLErrors, networkError, operation }) => {
-  if (process.env.NODE_ENV !== 'production') {
-    if (graphQLErrors)
-      console.error(
-        `[GraphQL errors]: ${graphQLErrors
-          .map(({ message, path }) => `${message} ${path?.join(',')}`)
-          .join(', ')}`,
-      )
+  if (graphQLErrors)
+    console.error(
+      `[GraphQL errors]: ${graphQLErrors
+        .map(({ message, path }) => `${message} ${path?.join(',')}`)
+        .join(', ')}`,
+    )
 
-    if (networkError)
-      console.error(`[Graphql error]: ${networkError}`)
+  if (networkError)
+    console.error(`[Graphql error]: ${networkError}`)
 
-      // eslint-disable-next-line @typescript-eslint/no-floating-promises
-    ;(async () => {
-      const graphql = await import('graphql')
+    // eslint-disable-next-line @typescript-eslint/no-floating-promises
+  ;(async () => {
+    const graphql = await import('graphql')
 
-      const gqlString = graphql
-        .print(operation.query)
-        .toString()
-        .replace(/(\r\n|\n|\r)/gm, ' ')
-        .replace(/\s\s+/g, ' ')
+    const gqlString = graphql
+      .print(operation.query)
+      .toString()
+      .replace(/(\r\n|\n|\r)/gm, ' ')
+      .replace(/\s\s+/g, ' ')
 
-      console.error(`[GraphQL operation]: ${gqlString}`)
-      console.error(`[GraphQL variables]: ${JSON.stringify(operation.variables)}`)
-    })()
-  }
+    console.error(`[GraphQL operation]: ${gqlString}`)
+    console.error(`[GraphQL variables]: ${JSON.stringify(operation.variables)}`)
+  })()
 })