+++ packages/next-ui/hooks/useDateTimeFormat.ts
@@ -1,0 +1,10 @@
+import { useRouter } from 'next/router'
+import { useMemo } from 'react'
+
+export type DateTimeFormatProps = Intl.DateTimeFormatOptions
+
+export function useDateTimeFormat(props?: DateTimeFormatProps) {
+  const { locale } = useRouter()
+  const formatter = useMemo(() => new Intl.DateTimeFormat(locale, props), [locale, props])
+  return formatter
+}