+++ packages/framer-scroller/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 
+## 2.1.19
+
+### Patch Changes
+
+- Updated dependencies [[`11bca2d2f`](https://github.com/graphcommerce-org/graphcommerce/commit/11bca2d2f7dbb7c5e2827c04eb0db43d4099f2fd)]:
+  - @graphcommerce/next-ui@4.11.1
+
+## 2.1.18
+
+### Patch Changes
+
+- Updated dependencies [[`9ec0338df`](https://github.com/graphcommerce-org/graphcommerce/commit/9ec0338dfe34d37b0f2c24e36ffa6ed13ea1145e)]:
+  - @graphcommerce/next-ui@4.11.0
+
 ## 2.1.17
 
 ### Patch Changes