+++ packages/magento-cart-checkout/CHANGELOG.md
@@ -1,5 +1,40 @@
 # Change Log
 
+## 3.0.33
+
+### Patch Changes
+
+- Updated dependencies [[`11bca2d2f`](https://github.com/graphcommerce-org/graphcommerce/commit/11bca2d2f7dbb7c5e2827c04eb0db43d4099f2fd)]:
+  - @graphcommerce/next-ui@4.11.1
+  - @graphcommerce/magento-cart@4.4.7
+  - @graphcommerce/magento-cart-coupon@3.1.1
+  - @graphcommerce/magento-cart-items@3.0.34
+  - @graphcommerce/magento-customer@4.6.2
+  - @graphcommerce/magento-store@4.2.13
+
+## 3.0.32
+
+### Patch Changes
+
+- Updated dependencies [[`d140fca34`](https://github.com/graphcommerce-org/graphcommerce/commit/d140fca3463b73e761e23fd1d9216305727f0c1a), [`c80e10d96`](https://github.com/graphcommerce-org/graphcommerce/commit/c80e10d96de18de73e7768f93aae806ce9cc45de), [`9ec0338df`](https://github.com/graphcommerce-org/graphcommerce/commit/9ec0338dfe34d37b0f2c24e36ffa6ed13ea1145e), [`735b78672`](https://github.com/graphcommerce-org/graphcommerce/commit/735b786724d5401cbe6e88f2515e121a1a0945b2)]:
+  - @graphcommerce/magento-cart@4.4.6
+  - @graphcommerce/magento-cart-coupon@3.1.0
+  - @graphcommerce/next-ui@4.11.0
+  - @graphcommerce/magento-store@4.2.12
+  - @graphcommerce/graphql@3.3.0
+  - @graphcommerce/magento-cart-items@3.0.33
+  - @graphcommerce/magento-customer@4.6.1
+
+## 3.0.31
+
+### Patch Changes
+
+- Updated dependencies [[`98ff2334d`](https://github.com/graphcommerce-org/graphcommerce/commit/98ff2334d1b7dedb8bc56ebe6abb50836eefedd3)]:
+  - @graphcommerce/magento-customer@4.6.0
+  - @graphcommerce/magento-cart@4.4.5
+  - @graphcommerce/magento-cart-items@3.0.32
+  - @graphcommerce/magento-cart-coupon@3.0.31
+
 ## 3.0.30
 
 ### Patch Changes